+++ src/lib/converter/types.ts
@@ -821,7 +821,7 @@ const referenceConverter: TypeConverter<
             name,
         );
 
-        if (type.flags & ts.TypeFlags.Substitution) {
+        if ((type.flags & ts.TypeFlags.Substitution) && name === "NoInfer" && ref.package === "typescript") {
             // NoInfer<T>
             ref.typeArguments = [
                 convertType(context, (type as ts.SubstitutionType).baseType),