+++ src/test/converter2/issues/gh3071.ts
@@ -1,0 +1,2 @@
+export interface Other {}
+export declare function m<Content>(content: Content): Content extends object ? Content : Other;